requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==5.0.6
django-allauth==64.1.0
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-imgix==1.1.0
django-rq==2.10.2
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
environ==1.0
git-filter-repo==2.45.0
gunicorn==23.0.0
idna==3.7
imgix==4.0.0
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
rq==1.16.2
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
